+++ src/main/java/nextstep/qna/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package nextstep.qna;
+package nextstep.qna.exception;
 
 public class ForbiddenException extends RuntimeException{
     public ForbiddenException() {
