+++ src/test/java/nextstep/qna/service/QnaServiceTest.java
@@ -1,6 +1,6 @@
 package nextstep.qna.service;
 
-import nextstep.qna.CannotDeleteException;
+import nextstep.qna.exception.CannotDeleteException;
 import nextstep.qna.domain.*;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.BeforeEach;
